+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.8.6
+
+* added `time_us`