+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.8.6
 
 * added `time_us`
+* updated `stringify` to correctly handle None-cases