+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-# v2.8.7
-
-* added setdefault and get to `DictObject`
+# v2.8.8